pyproject.toml
[project]
name = "receipt-management-api"
version = "0.1.0"
description = "REST API for receipt management"
authors = [
    {name = "Yuriy"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sqlalchemy[asyncio]>=2.0.23",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "pytest-cov>=4.1.0",
    "aiosqlite>=0.19.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests"
]
